requirements.txt
fastapi
uvicorn
face_recognition
pillow
numpy
chromadb
python-multipart
